requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
cryptography==42.0.5
PyJWT==2.8.0
